LayoutTests/http/tests/media/media-source/video-media-source-reject-append-after-reopening-expected.txt

Verify that append() calls on old SourceBuffers always fail when the MediaSource is closed and reopened.

EVENT(webkitsourceopen)
Adding SourceBuffers.
Triggering MediaSource to close and reopen.
EVENT(webkitsourceclose)
EVENT(webkitsourceopen)
Attempting to append to the old SourceBuffers.
Got an exception while appending: InvalidStateError
Got an exception while appending: InvalidStateError
END OF TEST
